Fixed MSVC Unit Tests + MISC Bugfixes
Turns out MSVC unit tests were failing for good goddamn reasons (mostly).

1) Had to handle uint32_t in GBL_TEST_COMPARE() macros explicitly for GBL_RESULT enum to work
2) uncovered alloca corruption in GblQuark paging tests
3) uncovered alloca corruption in GblStringView_rfind()
4) uncovered generic stack corruption in GblArrayMap tests

+++ test/source/containers/gimbal_array_map_test_suite.c
@@ -192,7 +192,7 @@ static GBL_RESULT GblArrayMapTestSuite_validate_variant_(GblTestSuite* pSelf, Gb
     GBL_TEST_COMPARE(GblArrayMap_getValue(&pMap, key), value);
     pVariant = GblArrayMap_getVariant(&pMap, key);
     GBL_TEST_VERIFY(pVariant);
-    uint32_t val = 0;
+    uintptr_t val = 0;
     GBL_TEST_COMPARE(GblVariant_getValuePeek(pVariant, &val), GBL_RESULT_SUCCESS);
     GBL_TEST_COMPARE(val, value);
     GBL_TEST_COMPARE(GblArrayMap_atValue(&pMap, key), value);
